Moved contextsigner out of lnwallet
